+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## To be Released
 
+- deprecation(constructor) Add deprecation command for `New()` method
+- feat(otel) Add opentelemetry instrumentation by default for all requests served by the server
+
 ## v1.9.2
 
 - fix(profiling_router): correctly initialize the empty router